Plenty fixes and optimizations
 Conflicts:
	src/pocketmine/block/AcaciaDoor.php
	src/pocketmine/block/Beacon.php
	src/pocketmine/block/BirchDoor.php
	src/pocketmine/block/BrewingStand.php
	src/pocketmine/block/Cauldron.php
	src/pocketmine/block/ChorusFlower.php
	src/pocketmine/block/ChorusPlant.php
	src/pocketmine/block/DarkOakDoor.php
	src/pocketmine/block/DragonEgg.php
	src/pocketmine/block/JungleDoor.php
	src/pocketmine/block/MobHead.php
	src/pocketmine/block/Sponge.php
	src/pocketmine/block/SpruceDoor.php
	src/pocketmine/block/StickyPiston.php
	src/pocketmine/command/defaults/SummonCommand.php
	src/pocketmine/entity/Cow.php
	src/pocketmine/entity/Donkey.php
	src/pocketmine/entity/ElderGuardian.php
	src/pocketmine/entity/Guardian.php
	src/pocketmine/entity/LingeringPotion.php
	src/pocketmine/entity/Mule.php
	src/pocketmine/entity/Painting.php
	src/pocketmine/entity/Pig.php
	src/pocketmine/entity/SkeletonHorse.php
	src/pocketmine/entity/Snake.php
	src/pocketmine/entity/ThrownEnderPearl.php
	src/pocketmine/entity/ThrownExpBottle.php
	src/pocketmine/entity/ZombieHorse.php
	src/pocketmine/entity/ZombieVillager.php
	src/pocketmine/event/block/ItemFrameDropItemEvent.php
	src/pocketmine/inventory/BeaconInventory.php
	src/pocketmine/inventory/CraftingManager.php
	src/pocketmine/inventory/EnderChestInventory.php
	src/pocketmine/item/Armor.php
	src/pocketmine/item/Boat.php
	src/pocketmine/item/EndCrystal.php
	src/pocketmine/item/Painting.php
	src/pocketmine/item/Potion.php
	src/pocketmine/item/PrismarineCrystal.php
	src/pocketmine/level/particle/SpellParticle.php
	src/pocketmine/level/sound/NoteblockSound.php
	src/pocketmine/tile/Beacon.php
	src/pocketmine/tile/BrewingStand.php
	src/pocketmine/tile/Cauldron.php
	src/pocketmine/tile/EnderChest.php
	src/pocketmine/tile/PistonArm.php
	src/pocketmine/tile/Skull.php
	src/pocketmine/utils/Color.php

+++ src/pocketmine/item/Potion.php
@@ -26,6 +26,7 @@ namespace pocketmine\item;
 use pocketmine\entity\Effect;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Human;
+
 class Potion extends Food{
 	const WATER_BOTTLE = 0;
 	const AWKWARD = 4;
@@ -187,7 +188,7 @@ class Potion extends Food{
 				$effect[] = Effect::getEffect(Effect::JUMP)->setAmplifier(0)->setDuration(8 * 60 * 20);
 				break;
 			case Potion::LEAPING_TWO:
-				$effect[] = Effect::getEffect(Effect::JUMP)->setAmplifier(1)->setDuration(1.5 * 60 * 20);
+				$effect[] = Effect::getEffect(Effect::JUMP)->setAmplifier(1)->setDuration((int) 1.5 * 60 * 20);
 				break;
 			case Potion::FIRE_RESISTANCE:
 				$effect[] = Effect::getEffect(Effect::FIRE_RESISTANCE)->setAmplifier(0)->setDuration(3 * 60 * 20);
@@ -202,7 +203,7 @@ class Potion extends Food{
 				$effect[] = Effect::getEffect(Effect::SPEED)->setAmplifier(0)->setDuration(8 * 60 * 20);
 				break;
 			case Potion::SPEED_TWO:
-				$effect[] = Effect::getEffect(Effect::SPEED)->setAmplifier(1)->setDuration(1.5 * 60 * 20);
+				$effect[] = Effect::getEffect(Effect::SPEED)->setAmplifier(1)->setDuration((int) 1.5 * 60 * 20);
 				break;
 			case Potion::SLOWNESS:
 				$effect[] = Effect::getEffect(Effect::SLOWNESS)->setAmplifier(0)->setDuration(1 * 60 * 20);
@@ -241,10 +242,10 @@ class Potion extends Food{
 				$effect[] = Effect::getEffect(Effect::STRENGTH)->setAmplifier(0)->setDuration(8 * 60 * 20);
 				break;
 			case Potion::STRENGTH_TWO:
-				$effect[] = Effect::getEffect(Effect::STRENGTH)->setAmplifier(1)->setDuration(1.5 * 60 * 20);
+				$effect[] = Effect::getEffect(Effect::STRENGTH)->setAmplifier(1)->setDuration((int) 1.5 * 60 * 20);
 				break;
 			case Potion::WEAKNESS:
-				$effect[] = Effect::getEffect(Effect::WEAKNESS)->setAmplifier(0)->setDuration(1.5 * 60 * 20);
+				$effect[] = Effect::getEffect(Effect::WEAKNESS)->setAmplifier(0)->setDuration((int) 1.5 * 60 * 20);
 				break;
 			case Potion::WEAKNESS_T:
 				$effect[] = Effect::getEffect(Effect::WEAKNESS)->setAmplifier(0)->setDuration(4 * 60 * 20);
@@ -267,7 +268,7 @@ class Potion extends Food{
 		}
 		return $effect;
 	}
-	
+
 	public function canBeConsumed() : bool{
 		return true;
 	}
@@ -287,7 +288,7 @@ class Potion extends Food{
 	public function canBeConsumedBy(Entity $entity) : bool{
 		return $entity instanceof Human;
 	}
-	
+
 	public function getMaxStackSize() {
 		return 1;
 	}
